pyproject.toml
[project]
name = "open-agent"
version = "0.1.0"
description = "Open Agent - A flexible AI agent framework with tools and MCP support"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Open Agent Team"}
]
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.5.0",
    "prompt-toolkit>=3.0.0",
    "anthropic>=0.39.0",
    "openai>=1.57.4",
    "agent-client-protocol>=0.6.0",
    "pillow>=12.1.1",
    # Web UI dependencies
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    # System tray dependencies
    "pystray>=0.19.0",
]

[project.scripts]
open-agent = "open_agent.cli:main"
open-agent-acp = "open_agent.acp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["open_agent*"]
exclude = ["tests*"]

# Include skills directory (git submodule) as package data
# After moving submodule to open_agent/skills
[tool.setuptools.package-data]
open_agent = ["skills/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
cache_dir = "workspace/.pytest_cache"
asyncio_mode = "auto"

[tool.pylint.messages_control]
# Disable warnings for Tool.execute method signature differences
# We intentionally use explicit parameters in subclasses for better type hints
disable = [
    "arguments-differ",  # Allow subclasses to have different parameter signatures
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]
